+++ data-raw/prep_UK2000.R
@@ -57,14 +57,10 @@ UKEnergy2000mats <- UKEnergy2000tidy %>%
 
   dplyr::mutate(
     # Create full U matrix
-    U = matsbyname::sum_byname(U_excl_EIOU, U_EIOU),
+    U = matsbyname::sum_byname(U_feed, U_EIOU),
     r_EIOU = matsbyname::quotient_byname(U_EIOU, U),
     r_EIOU = matsbyname::replaceNaN_byname(r_EIOU, val = 0)
   ) %>%
-  dplyr::rename(
-    U_feed = U_excl_EIOU
-  ) %>%
-  # dplyr::select(-U_EIOU, -U_excl_EIOU) %>%
   # Add S_units matrices
   dplyr::left_join(S_units, by = c("Country", "Year", "Energy.type", "Last.stage")) %>%
   tidyr::gather(key = matrix.name, value = matrix, R, U, U_EIOU, U_feed, V, Y, r_EIOU, S_units)