+++ NEWS.md
@@ -1,3 +1,99 @@
+# Recca 0.1.32 (2021-03-16)
+
+* Modified aggregation functions to use simpler (but verbose)
+  `if` statements to choose aggregation functions.
+  This change works around an error when calling the aggregation functions
+  from other packages.
+* New tests for new code paths.
+    * Now at 435 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# Recca 0.1.31 (2021-02-25)
+
+* New argument to aggregation functions: `pattern_type`
+  which tells how to match primary and final demand sector names for aggregation.
+* New tests for new feature.
+    * Now 421 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# Recca 0.1.30 (2021-02-22)
+
+* Fixed bugs in `finaldemand_aggregates()` and `finaldemand_aggregates_with_units()`
+  where argument `fd_sectors` was not respected for gross energy aggregation.
+* New and revamped tests to guard against regression.
+    * Now 417 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# Recca 0.1.29 (2021-02-01)
+
+* A `method_q_calculation` argument has been added to
+  the `calc_io_mats()` function.
+  This enables to calculate the q vector using either
+  a consumption-side or supply-side perspective.
+
+
+# Recca 0.1.28 (2020-12-23)
+
+* New function `find_p_industry_names()` looks at `R`, `V`, and `Y` matrices 
+  to find names of primary industries (industries that produce primary energy).
+  The function looks for prefixes among row (`R` and `V`) and column (`Y`) names
+  to identify primary industries.
+* New tests for new function.
+    * Now 416 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# Recca 0.1.27 (2020-12-08)
+
+* Move IEA aggregation functions out of this package and into the `IEATools` package.
+* Now using package constants for names of aggregate columns.
+  Column names match similar names in `IEATools`.
+* Now using better function-matching and -selecting code in `finaldemand_aggregates()`
+* Fewer tests due to moving some functions out of the package.
+    * Now 411 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# Recca 0.1.26 (2020-12-06)
+
+* `primary_aggregates()`, `finaldemand_aggregates(), and `finaldemand_aggregates_with_units()`
+  now require a column of `p_industries` or `fd_sectors` instead of a vector 
+  when a `.sutdata` is a data frame.
+  This change works around a problem with a common use case where 
+  the caller would create the `fd_sectors` column from the column names of `Y` vectors 
+  in the data frame.
+* Working to modernize the Recca code.
+    * Use R markdown in documentation (`` `term` `` instead of `\code{term}`).
+    * Use package constants for column name arguments 
+      (`Recca::sankey_cols$sankey` instead of "Sankey").
+* New tests for the new feature.
+    * Now up to 413 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# Recca 0.1.25 (2020-09-17)
+
+* New author Emmanuel Aramendia. Welcome!
+* Added tests for new features and old bugs.
+    * Now up to 407 tests, all of which pass.
+    * Test coverage remains at 100 %.
+* Test coverage now up to 100%.
+* Created a new function `calc_embodied_EIOU()` that calculates different embodied EIOU matrices for a given final demand
+* Created a new function `calc_erois()` that calculates different vector of product-level and industry-level EROIs 
+  using different assumptions.
+* Created a new function `calc_E_EIOU()` that calculates the `E_EIOU` matrix of EIOU energy use 
+  by unit of industry output, split by product.
+  It also calculates the direct energy extension vector e_EIOU of total EIOU energy use
+  by unit of industry output.
+* Fixed a bug in one of the vignettes related to the move from `U_excl_EIOU` to `U_feed`.
+* `prep_UK2000.R` and `prep_perfectsub.R` scripts
+  fixed to work with `IEATools`, which now uses
+  `U_feed` instead of `U_excl_EIOU`.
+
+
 # Recca 0.1.24 (2020-03-17)
 
 * New tests for new features and old bugs.